moved more things, but 'enableNFC_dialog' stops working -.-
